LP-2275 Resolve cyclic import issue
