+++ openedx/features/badging/helpers/__init__.py
@@ -1,0 +1,3 @@
+"""
+Helper methods for badging app
+"""
