+++ openedx/features/badging/helpers/badges.py
@@ -4,17 +4,14 @@ Helper methods
 import json
 from collections import OrderedDict
 
-from django.template.loader import render_to_string
 from django.urls import reverse
-from edx_notifications.data import NotificationMessage
-from edx_notifications.lib.publisher import get_notification_type, publish_notification_to_user
 
 from lms.djangoapps.courseware.courses import get_course_by_id
 from lms.djangoapps.teams import is_feature_enabled as is_teams_feature_enabled
 from lms.djangoapps.teams.models import CourseTeam
 from nodebb.constants import CONVERSATIONALIST_ENTRY_INDEX, TEAM_PLAYER_ENTRY_INDEX
 
-from .constants import (
+from openedx.features.badging.constants import (
     BADGES_KEY,
     BADGES_DATE_EARNED_KEY,
     BADGES_PROGRESS_KEY,
@@ -22,7 +19,6 @@ from .constants import (
     CONVERSATIONALIST,
     DISCUSSION_ID_KEY,
     DISCUSSION_COUNT_KEY,
-    EARNED_BADGE_NOTIFICATION_TYPE,
     FILTER_BADGES_ERROR,
     POST_COUNT_KEY,
     TEAM_COUNT_KEY,
@@ -32,7 +28,7 @@ from .constants import (
     THRESHOLD_LABEL_KEY,
     USERNAME_KEY
 )
-from .models import Badge
+from openedx.features.badging.models import Badge
 
 BADGE_TYPE_TEAM = TEAM_PLAYER[TEAM_PLAYER_ENTRY_INDEX]
 BADGE_TYPE_CONVERSATIONALIST = CONVERSATIONALIST[CONVERSATIONALIST_ENTRY_INDEX]
@@ -186,31 +182,6 @@ def get_badge_progress(index, badges, team_joined=True):
     elif not previous_badge or BADGES_DATE_EARNED_KEY in previous_badge:
         badge_progress = ('in-progress', 'In Progress')
     return badge_progress
-
-
-def send_user_badge_notification(user, my_badge_url, badge_name):
-    """
-    Send user new badge notification
-    :param user: User receiving the Notification
-    :param my_badge_url: Redirect url to my_badge view on notification click
-    :param badge_name: Newly earned badge
-    """
-    context = {
-        'badge_name': badge_name
-    }
-
-    body_short = render_to_string('philu_notifications/templates/user_badge_earned.html', context)
-
-    message = NotificationMessage(
-        msg_type=get_notification_type(EARNED_BADGE_NOTIFICATION_TYPE),
-        payload={
-            'from_user': user.username,
-            'path': my_badge_url,
-            'bodyShort': body_short,
-        }
-    )
-
-    publish_notification_to_user(user.id, message)
 
 
 def get_discussion_team_ids(course_id, discussion_room_id, badges):
